+++ migrations/server/postgres/2024-01-16-003004_add_qobuz_bundles_table/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS qobuz_bundles (
-    "id" INTEGER PRIMARY KEY NOT NULL,
+    "id" BIGSERIAL PRIMARY KEY NOT NULL,
     "bundle_version" VARCHAR(32) NOT NULL,
     "app_id" VARCHAR(32) NOT NULL,
     created TIMESTAMP NOT NULL DEFAULT NOW(),
