+++ migrations/server/postgres/2024-01-26-200343_fix_qobuz_id_on_albums_table/down.sql
@@ -1,6 +1,6 @@
-ALTER TABLE albums ADD COLUMN old_qobuz_id INTEGER DEFAULT NULL;
-UPDATE albums SET old_qobuz_id = CAST(qobuz_id as INTEGER);
+ALTER TABLE albums ADD COLUMN old_qobuz_id BIGINT DEFAULT NULL;
+UPDATE albums SET old_qobuz_id = CAST(qobuz_id as BIGINT);
 ALTER TABLE albums DROP COLUMN qobuz_id;
-ALTER TABLE albums ADD COLUMN qobuz_id INTEGER DEFAULT NULL;
+ALTER TABLE albums ADD COLUMN qobuz_id BIGINT DEFAULT NULL;
 UPDATE albums SET qobuz_id = old_qobuz_id;
 ALTER TABLE albums DROP COLUMN old_qobuz_id;
