+++ migrations/server/postgres/2024-02-06-014848_drop_not_null_constraint_from_track_sizes_bytes/up.sql
@@ -1,7 +1,7 @@
-ALTER TABLE track_sizes ADD COLUMN new_bytes INTEGER DEFAULT NULL;
+ALTER TABLE track_sizes ADD COLUMN new_bytes BIGINT DEFAULT NULL;
 UPDATE track_sizes SET new_bytes = bytes;
 ALTER TABLE track_sizes DROP COLUMN bytes;
-ALTER TABLE track_sizes ADD COLUMN bytes INTEGER DEFAULT NULL;
+ALTER TABLE track_sizes ADD COLUMN bytes BIGINT DEFAULT NULL;
 UPDATE track_sizes SET bytes = new_bytes;
 ALTER TABLE track_sizes DROP COLUMN new_bytes;
